backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-telegram-bot>=20.0
apscheduler>=3.10.0
p115client>=0.0.0.0
httpx>=0.27.0
aiofiles>=23.0.0
requests>=2.31.0
tmdbv3api>=1.7.7
pyyaml>=6.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2
openai>=1.0.0
a2wsgi>=1.10.0
sqlitetools
fake-useragent
wsgidav==4.0.0
cheroot
cryptography
aiofiles
python-multipart
pyotp
httpx
